# refactor(streams/to_text): use streaming for improved TextDecoder perf