This simple change leverages the `stream` option in the second parameter of `TextDecoder.decode`, to improve performance when calling the decode method multiple times in a row. It also adds an empty `.decode()` call outside of the loop to properly reset the internal buffer state of the decoder.

Thanks for your time!