@iuioiua 2024-04-02
Hi @nberlette, I wrote up a quick and dirty benchmark, and it shows that this change actually makes `toText()` more than 60% slower. I'll close this for now. However, please know that we'd always be happy to look at a PR that provides performance improvements and benchmarks. Either way, thank you for your PR.

Source:
```ts
import { toText } from "./streams/to_text.ts";

Deno.bench("x", async (b) => {
  const encoder = new TextEncoder();
  const text =
    "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.";
  const stream = ReadableStream.from([
    encoder.encode(text),
    encoder.encode(text),
    encoder.encode(text),
  ]);
  b.start();
  await toText(stream);
  b.end();
});
```
Main:
```
asher@Ashers-MacBook-Air deno_std % deno bench bench.ts
cpu: Apple M2
runtime: deno 1.42.1+240b362 (aarch64-apple-darwin)

file:///Users/asher/GitHub/deno_std/bench.ts
benchmark      time (avg)        iter/s             (min … max)       p75       p99      p995
--------------------------------------------------------------- -----------------------------
x               2.33 µs/iter     429,000.4     (1.92 µs … 1.13 ms) 2.21 µs 3.29 µs 4.58 µs
```

This PR:
```
asher@Ashers-MacBook-Air deno_std % deno bench bench.ts
cpu: Apple M2
runtime: deno 1.42.1+240b362 (aarch64-apple-darwin)

file:///Users/asher/GitHub/deno_std/bench.ts
benchmark      time (avg)        iter/s             (min … max)       p75       p99      p995
--------------------------------------------------------------- -----------------------------
x               3.83 µs/iter     260,960.3   (3.04 µs … 196.71 µs) 3.38 µs 7.21 µs 8.67 µs
```